+++ packages/core/test/test-util/test-services/v2/test-service/foo.ts
@@ -1,0 +1,5 @@
+import { TestEntityRequestBuilder } from './TestEntityRequestBuilder';
+
+export function foo(){
+  return new TestEntityRequestBuilder()
+}