+++ CHANGELOG.md
@@ -24,11 +24,11 @@
 
 ## Improvements
 
--
+- Reduce the number of circular dependencies within the SDK Core.
 
 ## Fixed Issues
 
--
+- 
 
 
 # 1.22.0